requirements.txt
streamlit
langchain-google-genai
langchain_core
python-dotenv
nltk
webdataset
streamlit-mic-recorder
gTTS
pydub
transformers
phonemizer
Levenshtein
python-ffmpeg